backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-py>=0.5.3",
    "fastapi[standard]>=0.115.8",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-openai>=0.3.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "supabase>=2.13.0",
]
